+++ board-frontend/src/scripts/Document.js
@@ -13,6 +13,7 @@ import {CanvasObjectType} from "./canvas/CanvasObjectType.js";
 import CanvasObjectWrapper from "./canvas/CanvasObjectWrapper.js";
 import SocketEventType from "./protocol/SocketEventType.js";
 import ClientMessageGetInvite from "./protocol/client/messages/ClientMessageGetInvite.js";
+import {apiUrl} from "./main.js";
 
 const documents = new Map();
 export let activeDocument = null;
@@ -47,6 +48,12 @@ class Document {
         //window.history.pushState(document.name, document.title, '/d/' + this.id); todo
         this.canvas.draw(-1);
         //this.addClient(localClient);
+        let request = new XMLHttpRequest();
+        request.addEventListener('load', (response) => {
+            console.log(response);
+        });
+        request.open('GET', apiUrl + '/document/' + this.id);
+        request.send();
     }
 
     close() {
@@ -119,14 +126,10 @@ socket.addMessageListener(ServerMessageType.UPDATE_CANVAS, (serverMessageUpdateC
     }
 });
 socket.addMessageListener(ServerMessageType.UPDATE_DOCUMENT, (serverMessageUpdateDocument) => {
-    console.log(serverMessageUpdateDocument.shared);
     documents.set(serverMessageUpdateDocument.id,
         new Document(serverMessageUpdateDocument.name + (serverMessageUpdateDocument.shared ? "(shared)" : ""),
             serverMessageUpdateDocument.id));
 });
-socket.addMessageListener(ServerMessageType.OPEN_DOCUMENT, (serverMessageOpenDocument) => {
-    activeDocument.canvas = serverMessageOpenDocument.canvas;
-})
 socket.addEventListener(SocketEventType.OPEN, () => {
     let invite;
     let index = document.location.href.lastIndexOf('invite=');