+++ board-frontend/src/scripts/canvas/canvasObjects/CanvasImage.js
@@ -1,18 +1,12 @@
 import CanvasObject from "../CanvasObject.js";
 import {ctx} from "../Canvas.js";
+import {apiUrl} from "../../main.js";
 
 export default class CanvasImage extends CanvasObject {
     constructor(reader) {
         super(reader);
-        this.type = reader.readUint8();
-        switch (this.type) {
-            case CanvasImageType.PNG:
-                this.image = document.createElement('img');
-                this.image.src = 'data:image/png;base64,' + reader.readBase64();
-                break;
-            default:
-                console.warn('unsupported image type ' + this.type);
-        }
+        this.image = document.createElement('img');
+        this.image.src = apiUrl + '/resources/canvasImage/' + reader.readBigInt64();//todo url
     }
 
     draw() {
@@ -29,13 +23,7 @@ export default class CanvasImage extends CanvasObject {
         let shape = Object.create(this.prototype);
         shape.x = x;
         shape.y = y;
-        shape.type = CanvasImageType.PNG;
         shape.image = image;
         return shape;
     }
-}
-
-export const CanvasImageType = {
-    PNG:0,
-    URL:1
 }