+++ board-frontend/src/scripts/main.js
@@ -1,2 +1,4 @@
 import './Document.js';
 import './FileUpload.js'
+
+export const apiUrl = 'http://localhost/api';