+++ board-room/src/main/java/net/stzups/board/room/server/websocket/protocol/server/ServerMessageType.java
@@ -4,11 +4,10 @@ public enum ServerMessageType {
     ADD_CLIENT(0),
     REMOVE_CLIENT(1),
     UPDATE_CANVAS(2),
-    OPEN_DOCUMENT(3),
-    UPDATE_DOCUMENT(4),
-    ADD_USER(5),
-    DELETE_DOCUMENT(6),
-    GET_INVITE(7)
+    UPDATE_DOCUMENT(3),
+    ADD_USER(4),
+    DELETE_DOCUMENT(5),
+    GET_INVITE(6)
     ;
 
     private final int id;