+++ CHANGELOG.md
@@ -132,8 +132,9 @@ Bugfixes, better urls/file paths
 - Improve routing
 
 ## v0.5
-Canvas features
+Canvas features, add api for server resources like documents
 
 ### v0.5.0
-CanvasImage upload
+CanvasImage upload, add api
 - Ability to upload images on to the board
+- API: http://localhost/api/document/id